Move network_utils.
network_utils has a publicly exposed and documented method so we can't
move it without leaving a stub in place. Since the file is pretty simple
with a single exported method, I moved the whole file back.
