+++ google_compute_engine/networking/network_daemon.py
@@ -29,7 +29,7 @@ from google_compute_engine import constants
 from google_compute_engine import file_utils
 from google_compute_engine import logger
 from google_compute_engine import metadata_watcher
-from google_compute_engine.networking import network_utils
+from google_compute_engine import network_utils
 from google_compute_engine.networking.dhcp_lease_refresh import dhcp_lease_refresh
 from google_compute_engine.networking.ip_forwarding import ip_forwarding
 from google_compute_engine.networking.network_setup import network_setup
